Queue Implementation
The queue must be a circular doubly linked consisting of q-elements. A queue consists of a head-pointer and a set of q-elements. The first q-element in the queue is the first element in the queue. The head pointer points to the first element. See the figures for a better explanation. A q-element is a structure, consisting of a prev and next pointer, and a payload consisting of 1 integer.

The functions that you implement are:

1.     item = NewItem(); // returns a pointer to a new q-element, uses memory allocation

2.     head = newQueue() // creates a empty queue, that is the header pointer is set to null.

3.     AddQueue(head, item) // adds a queue item, pointed to by “item”, to the queue pointed to by head.

4.     item = DelQueue(head) // deletes an item from head and returns a pointer to the deleted item. If the queue is already empty, flag error.