Extending project-2,implemented routines that implements semaphores and then use them to implement the readers/writers problem.
Step 1:

Declare a semaphore type (name it Semaphore_t) . The semaphore is a counter and a queue of TCBs. Semaphores declared/used in the end program will be of this type.

Step 2:

Implement the routines

Semaphore_t CreateSem(int InputValue) // mallocs a semaphore structure, initializes it to the InitValue and returns the pointer to the semaphore.

void P(Semaphore_t * sem)  // takes a pointer to a semaphore and performs P, i.e. decrements the semaphore, and if the value is less than zero then blocks the thread in the queue associated with the semaphore.

void V(Semaphore_t * sem)  // increments the semaphore, and if the value is 0 or negative,
then takes a PCB out of the semaphore queue and puts it into the ReadyQ.
Note: The V routine also "yields" to the next runnable thread.

-- put the code in a file called “sem.h”. Of course you need to include threads.h into this file (to get access to yield) and threads includes tcb.h.
Step4:
Test it using readers writers problem