Developed a module that does message passing via ports.
Step 1:

Declare a port. A port is something that contains N messages, make N=10. There would be semaphores associated with the port, for controlling synchronization. The port “contains” messages hence a port may be a pointer to a Q of messages, or an array of messages.

Declare a message type. For simplicity, a message is an array of 10 integers.

Declare a set (array of ports). The ports are numbered 0 to 99.

Step 2:

Implement the routines

Send (with appropriate arguments). The send is a “blocking” asynchronous send that is: it places the message in the port and returns but blocks if port is full.

 

Receive (with appropriate arguments). The receive is a “blocking” receive.

 

Put the code in msgs.h, and include sem.h in the file msgs.h

 

Step 3:

Test the message passing system by writing one (or more) servers and having several clients talk to the servers.