Client-server architecture for file transfer. Clients can upload files and server accepts file-name and file data from client and writes them to the files. Server was single threaded and could support 3 concurrent transfers.
Overview:

For this project you are to write build clients and one server that does file transfers. The clients can only upload files, and the server accepts file-name and file data from clients and writes them to files.

All ports are static. The client sends a message to the server port with a “filename” (string) and the file server changes that string to “filename.server”. The client then sends the file data to the server and the server writes the data to the file “filename.server”. (Note: The client may have to tell server file transfer is over).

Details:

1.      File server can do file receives simultaneously from multiple clients. Make the max number of concurrent transfers 3. If a fourth client tries, it should be told “do not do that”.

2.      Maximum file size is 1MB (approx), server enforces this.

3.      Maximum file length used by client – 15 chars, server enforces this (filename)

4.      Clients and servers do not share memory or file systems.

5.      Each client sends one file and then stops (see details below)